Rename components and fix scrolling behaviour
