+++ app/routes/unterstuetzung/route.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { type MetaArgs } from "react-router";
 import { twJoin } from "tailwind-merge";
 
@@ -56,19 +56,6 @@ function SocialProofImage() {
 
 export default function Index() {
   const [isAppointmentsVisible, setIsAppointmentsVisible] = useState(false);
-
-  // Ensure page scrolls to the correct section when navigating to the page
-  // Brief delay to allow elements to render which shift the content
-  useEffect(() => {
-    const scrollTargetId = location.hash.substring(1);
-    if (scrollTargetId === "angebote" || scrollTargetId === "hilfe") {
-      const timerId = setTimeout(() => {
-        const element = document.getElementById(scrollTargetId);
-        element?.scrollIntoView({ behavior: "auto" });
-      }, 100);
-      return () => clearTimeout(timerId);
-    }
-  }, []);
 
   const iframeButtons = [
     {
